Fix -Wsign-compare warning in Touch_CST226::getTouchRaw
The loop counter was int while count is uint_fast8_t. On ARM cores
uint_fast8_t is unsigned int, so the comparison triggers
-Wsign-compare (on xtensa/riscv it is unsigned char, promoted to int,
which is why ESP32 builds never warned). Harmless at runtime; use the
same type as count.

+++ src/lgfx/v1/touch/Touch_CSTxxx.cpp
@@ -315,7 +315,7 @@ namespace lgfx
         if( requested_count > 1 ) // an array of touch_point_t was provided
         {
           int pIdx = 0; // points index with corrected offset
-          for (int i=0; i<count; i++) {
+          for (uint_fast8_t i=0; i<count; i++) {
             tp[i].id = i;
             tp[i].size = readdata[pIdx+4]; // pressure
             tp[i].x = (uint16_t)((readdata[pIdx+1] << 4) | ((readdata[pIdx+3] >> 4) & 0xf));
